[Core] Define Core and overhaul Executor signature
Summary:

Define the Core struct that will be the shared state object
between Future and Promise.

Overhaul the Executor interface due to the need of Core to
hold a borrow to an Executor trait, which only works if Executor
trait does not take Type Parameters.

Use State enum in the FSM struct.

Add some helper cache line alignment helpers.