[MicroSpinLock] Use compare_exchange in MicroSpinLock and add benchmark
Use the extended_compare_and_swap feature that allows compare_exchange use
which allows us to specify the MemoryOrdering in the case we successfully
swap or do not swap an Atomic. Also add benchmarks as well.

Some other rough cleanup for Try<T> => Result<T,E> and bubbling that up to
Core as well, will see how that goes.