[Core] Finish implementation of do_callback
Summary:

Finish the implementation of do_callback and change callback and result to
be UnsafeCells so that we can modify them.

Also add a little more functionality to AlignedAs, as well as properly exporting
the scope_exit! macro and make the ScopeGuard struct public so it can be used.

Test Plan: cargo test